+++ content/lib/pyswitch/controller/pager.py
@@ -12,11 +12,10 @@ class PagerAction(Callback, Action):
 
     # Proxy action which directly selects a page
     class _DirectPageProxy(Action):
-        def __init__(self, pager, page_index, use_leds, id, display, enable_callback):
+        def __init__(self, pager, page_index, use_leds, id, enable_callback):
             Action.__init__(self, {
                 "useSwitchLeds": use_leds,
                 "id": id,
-                "display": display,
                 "enableCallback": enable_callback
             })
             self.__pager = pager
@@ -24,6 +23,9 @@ class PagerAction(Callback, Action):
 
         # Called when the switch is pushed down
         def push(self):
+            if not hasattr(self.__pager, "switch") or not self.__pager.switch:
+                raise Exception("Every Pager must be assigned to a switch. See Select Page action.")
+            
             if not self.__pager.pages: 
                 return
             
@@ -36,18 +38,24 @@ class PagerAction(Callback, Action):
             self.__pager.update_displays()
 
         def update_displays(self):
+            if not hasattr(self, "switch"):
+                return 
+            
             if self.__page_index == None:
                 return
             
-            if len(self.__pager.pages) <= self.__page_index - 1:
-                return
-
             if not self.enabled:
                 return
             
-            page = self.__pager.pages[self.__page_index]
-            self.__pager.apply_page_to_action_displays(self, page, self.__page_index)
-
+            page_select = self.__pager.pages[self.__page_index]
+            
+            if "color" in page_select:
+                self.switch_color = page_select["color"]
+            else:
+                self.switch_color = (255, 255, 255)
+                
+            is_current = self.__page_index == self.__pager.current_page_index
+            self.switch_brightness = self.__pager.led_brightness_on if is_current else self.__pager.led_brightness_off              
 
     # The PagerAction is used to control multiple other actions to provide paging. Define this action for one switch which will by default rotate through the defined pages.
     # For the actions you want to be part of pages, just assign them using the paging buttons (which will set the id and enable_callback parameters for you).
@@ -57,7 +65,7 @@ class PagerAction(Callback, Action):
     # If you want to rotate through pages, just use this action with select_page set to None, which is the default. The switch will then rotate through the pages. 
     # 
     # <b>Directly Select Pages:</b>
-    # If you want to have one switch dedicated to select each page, set select_page here to the page you want to select with this switch, and use the "Select Page" action (assigned to this pager) for the other switches to select the further pages directly.
+    # If you want to have one switch dedicated to select each page, set select_page here to the page you want to select with the switch this pager is assigned to, and use the "Select Page" action (assigned to this pager) for the other switches to select the further pages directly.
     def __init__(self, 
                  pages,                         # This has to be a list of dicts like follows:
                                                 # {
@@ -72,8 +80,6 @@ class PagerAction(Callback, Action):
                  led_brightness = 0.15,         # LED brightness for the pager in range [0..1]. Only used when selet_page is None (rotate mode).
                  led_brightness_off = 0.02,     # LED brightness for the pager when select_page is set and the page is not currently selected. Range [0..1].
                  led_brightness_on = 0.3,       # LED brightness for the pager when select_page is set and the page is currently selected. Range [0..1].
-                 display_dim_factor_on = 1,     # Display dim factor used when in rotation mode or if the current page is select_page
-                 display_dim_factor_off = 0.2,  # Display dim factor used when select_page is selt and the current page is not select_page
                  mappings = [],                 # List of mappings the paging depends on (optional)
                  use_leds = True,
                  id = None,
@@ -92,8 +98,6 @@ class PagerAction(Callback, Action):
         self.led_brightness = led_brightness
         self.led_brightness_on = led_brightness_on
         self.led_brightness_off = led_brightness_off
-        self.display_dim_factor_on = display_dim_factor_on
-        self.display_dim_factor_off = display_dim_factor_off
         self.__proxies = []
 
         self.current_page_index = -1
@@ -102,21 +106,25 @@ class PagerAction(Callback, Action):
 
         self.enable_callback = PagerAction._EnableCallback(self)
 
-    # This controls a PagerAction from another switch, making it possible to directly select pages with dedicated switches.
+    # This controls a Pager Action from another switch, making it possible to directly select pages with dedicated switches.
     # 
-    # You always need to have one (or more) PagerAction(s) which defines the pages. For selecting the other pages of a pager, create "Select Page" actions on the other switches.
+    # You always need to have a Pager Action which defines the pages, which also MUST be assigned to a switch. Set "select_page" on this pager to the first page, and for selecting the other pages, create "Select Page" actions for other switches. For example, to have 3 pages and 3 switches:
+    # <ul>
+    #     <li>Switch 1: Pager Action with "select_page" set to the first page</li>
+    #     <li>Switch 2: Select Page action set to page 2</li>
+    #     <li>Switch 3: Select Page action set to page 3</li>
+    # </ul>
+    # The LED and display brightness settings are determined from the connected pager and can be set there if needed.
     def proxy(self, 
               page_id,                    # Sets the page to be selected with this action
               use_leds = True, 
               id = None, 
-              display = None, 
               enable_callback = None
         ):
         proxy = self._DirectPageProxy(self, 
                                       page_index = self._get_page_index(page_id),
                                       use_leds = use_leds,
                                       id = id,
-                                      display = display,
                                       enable_callback = enable_callback)
         
         self.__proxies.append(proxy)
@@ -167,39 +175,40 @@ class PagerAction(Callback, Action):
         if not self.enabled:
             return
         
-        page = self.pages[self.current_page_index if self.__select_page_index == None else self.__select_page_index]
-        self.apply_page_to_action_displays(self, page, self.__select_page_index)
+        page_current = self.pages[self.current_page_index]
+        
+        # LEDs
+        if self.__select_page_index != None:
+            # Direct select
+            page_select = self.pages[self.__select_page_index]
+            
+            if "color" in page_select:
+                self.switch_color = page_select["color"]
+            else:
+                self.switch_color = (255, 255, 255)
 
+            is_current = self.__select_page_index == self.current_page_index                
+            self.switch_brightness = self.led_brightness_on if is_current else self.led_brightness_off                     
+        else:
+            # Rotating
+            if "color" in page_current:                
+                self.switch_color = page_current["color"]
+            else:
+                self.switch_color = (255, 255, 255)
+            
+            self.switch_brightness = self.led_brightness
+
+        # Label (doing the same thing for rotary or direct modes)
+        if self.label:
+            self.label.text = page_current["text"] if "text" in page_current else ""
+            
+            if self.label.back_color:
+                if "color" in page_current:                
+                    self.label.back_color = page_current["color"]
+                else:
+                    self.label.back_color = (255, 255, 255)
+        
+        # Update all proxies, too
         for proxy in self.__proxies:
             proxy.update_displays()
 
-    # Used by this class and proxies
-    def apply_page_to_action_displays(self, action, page, page_index):
-        if not hasattr(action, "switch"):
-            return
-        
-        page_color = page["color"] if "color" in page else None
-
-        if page_color:
-            action.switch_color = page["color"]
-
-        if action.label:
-            action.label.text = page["text"] if "text" in page else ""
-            
-        if page_index == None:
-            action.switch_brightness = self.led_brightness
-
-            if action.label and action.label.back_color and page_color:
-                action.label.back_color = page_color
-        else:
-            is_current = page_index == self.current_page_index
-
-            action.switch_brightness = self.led_brightness_on if is_current else self.led_brightness_off     
-
-            if action.label and action.label.back_color and page_color:
-                factor = self.display_dim_factor_on if is_current else self.display_dim_factor_off
-                action.label.back_color = (
-                    int(page_color[0] * factor),
-                    int(page_color[1] * factor),
-                    int(page_color[2] * factor)
-                )