+++ test/pyswitch/test_action_pager.py
@@ -199,7 +199,7 @@ class TestPagerAction(unittest.TestCase):
         self.assertEqual(switch_pager.brightness, 0)
 
         
-    def test_with_label(self):
+    def test_with_label_and_leds(self):
         display = DisplayLabel(layout = {
             "font": "foo",
             "backColor": (0, 0, 0)
@@ -287,10 +287,77 @@ class TestPagerAction(unittest.TestCase):
         self.assertEqual(pager.label.back_color, (3, 4, 5))
 
 
-    def test_select_page(self):
+    def test_without_page_color(self):
+        pager = PagerAction(
+            pages = [
+                {
+                    "id": 1, 
+                    "color": (3, 4, 5),
+                    "text": "foo"
+                },
+                {
+                    "id": 2, 
+                    "text": "bar"
+                },
+            ],
+            led_brightness = 0.4
+        )
+
+        action_1 = MockAction({ "enableCallback": pager.enable_callback, "id": 1 })
+        action_2 = MockAction({ "enableCallback": pager.enable_callback, "id": 2 })
+
+        switch = MockFootswitch(
+            actions = [
+                action_1, 
+                action_2
+            ]
+        )
+        
+        switch_pager = MockFootswitch(
+            pixels = [0],
+            actions = [
+                pager
+            ]
+        )
+        
+        appl = MockController2(
+            inputs = [
+                switch, 
+                switch_pager
+            ]
+        )
+        
+        pager.init(appl, switch_pager)
+        pager.update_displays()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), True)
+        self.assertEqual(pager.enable_callback.enabled(action_2), False)
+
+        self.assertEqual(switch_pager.color, (3, 4, 5))
+        self.assertEqual(switch_pager.brightness, 0.4)
+
+        # Select page 2
+        pager.push()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), False)
+        self.assertEqual(pager.enable_callback.enabled(action_2), True)
+
+        self.assertEqual(switch_pager.color, (255, 255, 255))
+        self.assertEqual(switch_pager.brightness, 0.4)
+
+        # Select page 1 again
+        pager.push()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), True)
+        self.assertEqual(pager.enable_callback.enabled(action_2), False)
+
+        self.assertEqual(switch_pager.color, (3, 4, 5))
+        self.assertEqual(switch_pager.brightness, 0.4)
+
+
+    def test_display_no_back_color(self):
         display = DisplayLabel(layout = {
-            "font": "foo",
-            "backColor": (0, 0, 0)
+            "font": "foo"
         })
 
         pager = PagerAction(
@@ -307,11 +374,210 @@ class TestPagerAction(unittest.TestCase):
                 },
             ],
             display = display,
+            led_brightness = 0.4
+        )
+
+        action_1 = MockAction({ "enableCallback": pager.enable_callback, "id": 1 })
+        action_2 = MockAction({ "enableCallback": pager.enable_callback, "id": 2 })
+
+        switch = MockFootswitch(
+            actions = [
+                action_1, 
+                action_2
+            ]
+        )
+        
+        switch_pager = MockFootswitch(
+            pixels = [0],
+            actions = [
+                pager
+            ]
+        )
+        
+        appl = MockController2(
+            inputs = [
+                switch, 
+                switch_pager
+            ]
+        )
+        
+        # These calls shall do nothing before init!
+        pager.push()
+        pager.update_displays()
+
+        pager.init(appl, switch_pager)
+        pager.update_displays()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), True)
+        self.assertEqual(pager.enable_callback.enabled(action_2), False)
+
+        self.assertEqual(switch_pager.color, (3, 4, 5))
+        self.assertEqual(switch_pager.brightness, 0.4)
+
+        self.assertEqual(pager.label.text, "foo")
+        self.assertEqual(pager.label.back_color, None)
+
+        # Select page 2
+        pager.push()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), False)
+        self.assertEqual(pager.enable_callback.enabled(action_2), True)
+
+        self.assertEqual(switch_pager.color, (3, 6, 8))
+        self.assertEqual(switch_pager.brightness, 0.4)
+
+        self.assertEqual(pager.label.text, "bar")
+        self.assertEqual(pager.label.back_color, None)
+
+        # Select page 1 again
+        pager.push()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), True)
+        self.assertEqual(pager.enable_callback.enabled(action_2), False)
+
+        self.assertEqual(switch_pager.color, (3, 4, 5))
+        self.assertEqual(switch_pager.brightness, 0.4)
+
+        self.assertEqual(pager.label.text, "foo")
+        self.assertEqual(pager.label.back_color, None)
+
+    
+    ###########################################################################################################
+
+
+    def test_select_page(self):
+        display = DisplayLabel(layout = {
+            "font": "foo",
+            "backColor": (0, 0, 0)
+        })
+
+        pager = PagerAction(
+            pages = [
+                {
+                    "id": 1, 
+                    "color": (3, 4, 5),
+                    "text": "foo"
+                },
+                {
+                    "id": 2, 
+                    "text": "bar"
+                },
+            ],
+            display = display,
             led_brightness_on = 0.4,
             led_brightness_off = 0.3,
-            display_dim_factor_on = 1,
-            display_dim_factor_off = 0.5,
             select_page = 2
+        )
+
+        proxy = pager.proxy(
+            page_id = 1,
+            use_leds = True
+        )
+
+        action_1 = MockAction({ "enableCallback": pager.enable_callback, "id": 1 })
+        action_2 = MockAction({ "enableCallback": pager.enable_callback, "id": 2 })
+
+        switch = MockFootswitch(
+            actions = [
+                action_1, 
+                action_2
+            ]
+        )
+        
+        switch_pager = MockFootswitch(
+            pixels = [0],
+            actions = [
+                pager
+            ]
+        )
+        
+        appl = MockController2(
+            inputs = [
+                switch, 
+                switch_pager
+            ]
+        )
+        
+        # These calls shall do nothing before init!
+        pager.push()
+        pager.update_displays()
+
+        pager.init(appl, switch_pager)
+        pager.update_displays()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), True)
+        self.assertEqual(pager.enable_callback.enabled(action_2), False)
+
+        self.assertEqual(switch_pager.color, (255, 255, 255))
+        self.assertEqual(switch_pager.brightness, 0.3)
+
+        self.assertEqual(pager.label.text, "foo")
+        self.assertEqual(pager.label.back_color, (3, 4, 5))
+
+        # Select page 2
+        pager.push()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), False)
+        self.assertEqual(pager.enable_callback.enabled(action_2), True)
+
+        self.assertEqual(switch_pager.color, (255, 255, 255))
+        self.assertEqual(switch_pager.brightness, 0.4)
+
+        self.assertEqual(pager.label.text, "bar")
+        self.assertEqual(pager.label.back_color, (255, 255, 255))
+
+        # Select page 2 again
+        pager.push()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), False)
+        self.assertEqual(pager.enable_callback.enabled(action_2), True)
+
+        self.assertEqual(switch_pager.color, (255, 255, 255))
+        self.assertEqual(switch_pager.brightness, 0.4)
+
+        self.assertEqual(pager.label.text, "bar")
+        self.assertEqual(pager.label.back_color, (255, 255, 255))
+
+        # Use proxy to go back to page 1
+        proxy.push()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), True)
+        self.assertEqual(pager.enable_callback.enabled(action_2), False)
+
+        self.assertEqual(switch_pager.color, (255, 255, 255))
+        self.assertEqual(switch_pager.brightness, 0.3)
+
+        self.assertEqual(pager.label.text, "foo")
+        self.assertEqual(pager.label.back_color, (3, 4, 5))
+
+
+    def test_select_page_no_color_on_page_1(self):
+        display = DisplayLabel(layout = {
+            "font": "foo",
+            "backColor": (0, 0, 0)
+        })
+
+        pager = PagerAction(
+            pages = [
+                {
+                    "id": 1, 
+                    "text": "foo"
+                },
+                {
+                    "id": 2, 
+                    "color": (3, 6, 8),
+                    "text": "bar"
+                },
+            ],
+            display = display,
+            led_brightness_on = 0.4,
+            led_brightness_off = 0.3,
+            select_page = 2
+        )
+
+        proxy = pager.proxy(
+            page_id = 1,
+            use_leds = True
         )
 
         action_1 = MockAction({ "enableCallback": pager.enable_callback, "id": 1 })
@@ -351,8 +617,8 @@ class TestPagerAction(unittest.TestCase):
         self.assertEqual(switch_pager.color, (3, 6, 8))
         self.assertEqual(switch_pager.brightness, 0.3)
 
-        self.assertEqual(pager.label.text, "bar")
-        self.assertEqual(pager.label.back_color, (1, 3, 4))
+        self.assertEqual(pager.label.text, "foo")
+        self.assertEqual(pager.label.back_color, (255, 255, 255))
 
         # Select page 2
         pager.push()
@@ -377,6 +643,18 @@ class TestPagerAction(unittest.TestCase):
 
         self.assertEqual(pager.label.text, "bar")
         self.assertEqual(pager.label.back_color, (3, 6, 8))
+
+        # Use proxy to go back to page 1
+        proxy.push()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), True)
+        self.assertEqual(pager.enable_callback.enabled(action_2), False)
+
+        self.assertEqual(switch_pager.color, (3, 6, 8))
+        self.assertEqual(switch_pager.brightness, 0.3)
+
+        self.assertEqual(pager.label.text, "foo")
+        self.assertEqual(pager.label.back_color, (255, 255, 255))
 
 
     def test_no_pages(self):
@@ -442,6 +720,11 @@ class TestPagerAction(unittest.TestCase):
 
 
     def test_proxy(self):
+        display = DisplayLabel(layout = {
+            "font": "foo",
+            "backColor": (0, 0, 0)
+        })
+
         pager = PagerAction(
             pages = [
                 {
@@ -457,30 +740,17 @@ class TestPagerAction(unittest.TestCase):
             ],
             led_brightness_on = 0.4,
             led_brightness_off = 0.3,
-            display_dim_factor_on = 1,
-            display_dim_factor_off = 0.5
+            display = display
         )
-
-        display_1 = DisplayLabel(layout = {
-            "font": "foo",
-            "backColor": (0, 0, 0)
-        })
 
         proxy_1 = pager.proxy(
             page_id = 1,
-            use_leds = True,
-            display = display_1
+            use_leds = True
         )
-
-        display_2 = DisplayLabel(layout = {
-            "font": "foo",
-            "backColor": (0, 0, 0)
-        })
 
         proxy_2 = pager.proxy(
             page_id = 2,
-            use_leds = True,
-            display = display_2
+            use_leds = True
         )
 
         action_1 = MockAction({ "enableCallback": pager.enable_callback, "id": 1 })
@@ -539,10 +809,8 @@ class TestPagerAction(unittest.TestCase):
         self.assertEqual(switch_proxy_1.brightness, 0.4)
         self.assertEqual(switch_proxy_2.brightness, 0.3)
 
-        self.assertEqual(proxy_1.label.text, "foo")
-        self.assertEqual(proxy_2.label.text, "bar")
-        self.assertEqual(proxy_1.label.back_color, (3, 4, 5))
-        self.assertEqual(proxy_2.label.back_color, (1, 3, 4))
+        self.assertEqual(pager.label.text, "foo")
+        self.assertEqual(pager.label.back_color, (3, 4, 5))
 
         # Select page 2
         proxy_2.push()
@@ -555,10 +823,8 @@ class TestPagerAction(unittest.TestCase):
         self.assertEqual(switch_proxy_1.brightness, 0.3)
         self.assertEqual(switch_proxy_2.brightness, 0.4)
 
-        self.assertEqual(proxy_1.label.text, "foo")
-        self.assertEqual(proxy_2.label.text, "bar")
-        self.assertEqual(proxy_1.label.back_color, (1, 2, 2))
-        self.assertEqual(proxy_2.label.back_color, (3, 6, 8))
+        self.assertEqual(pager.label.text, "bar")
+        self.assertEqual(pager.label.back_color, (3, 6, 8))
 
         # Select page 1
         proxy_1.push()
@@ -571,18 +837,20 @@ class TestPagerAction(unittest.TestCase):
         self.assertEqual(switch_proxy_1.brightness, 0.4)
         self.assertEqual(switch_proxy_2.brightness, 0.3)
 
-        self.assertEqual(proxy_1.label.text, "foo")
-        self.assertEqual(proxy_2.label.text, "bar")
-        self.assertEqual(proxy_1.label.back_color, (3, 4, 5))
-        self.assertEqual(proxy_2.label.back_color, (1, 3, 4))
+        self.assertEqual(pager.label.text, "foo")
+        self.assertEqual(pager.label.back_color, (3, 4, 5))
 
-    
-    def test_proxy_invalid_page(self):
+
+    def test_proxy_no_color_on_page_1(self):
+        display = DisplayLabel(layout = {
+            "font": "foo",
+            "backColor": (0, 0, 0)
+        })
+
         pager = PagerAction(
             pages = [
                 {
                     "id": 1, 
-                    "color": (3, 4, 5),
                     "text": "foo"
                 },
                 {
@@ -593,12 +861,17 @@ class TestPagerAction(unittest.TestCase):
             ],
             led_brightness_on = 0.4,
             led_brightness_off = 0.3,
-            display_dim_factor_on = 1,
-            display_dim_factor_off = 0.5
+            display = display
         )
 
         proxy_1 = pager.proxy(
-            page_id = 30
+            page_id = 1,
+            use_leds = True
+        )
+
+        proxy_2 = pager.proxy(
+            page_id = 2,
+            use_leds = True
         )
 
         action_1 = MockAction({ "enableCallback": pager.enable_callback, "id": 1 })
@@ -625,25 +898,188 @@ class TestPagerAction(unittest.TestCase):
             ]
         )
 
+        switch_proxy_2 = MockFootswitch(
+            pixels = [0],
+            actions = [
+                proxy_2
+            ]
+        )
+
         appl = MockController2(
             inputs = [
                 switch, 
                 switch_pager,
-                switch_proxy_1
+                switch_proxy_1,
+                switch_proxy_2
             ]
         )
         
         pager.init(appl, switch_pager)
         proxy_1.init(appl, switch_proxy_1)
+        proxy_2.init(appl, switch_proxy_2)
 
         pager.update_displays()
         proxy_1.update_displays()
+        proxy_2.update_displays()
 
         self.assertEqual(pager.enable_callback.enabled(action_1), True)
         self.assertEqual(pager.enable_callback.enabled(action_2), False)
 
+        self.assertEqual(switch_proxy_1.color, (255, 255, 255))
+        self.assertEqual(switch_proxy_2.color, (3, 6, 8))
+        self.assertEqual(switch_proxy_1.brightness, 0.4)
+        self.assertEqual(switch_proxy_2.brightness, 0.3)
+
+        self.assertEqual(pager.label.text, "foo")
+        self.assertEqual(pager.label.back_color, (255, 255, 255))
+
         # Select page 2
+        proxy_2.push()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), False)
+        self.assertEqual(pager.enable_callback.enabled(action_2), True)
+
+        self.assertEqual(switch_proxy_1.color, (255, 255, 255))
+        self.assertEqual(switch_proxy_2.color, (3, 6, 8))
+        self.assertEqual(switch_proxy_1.brightness, 0.3)
+        self.assertEqual(switch_proxy_2.brightness, 0.4)
+
+        self.assertEqual(pager.label.text, "bar")
+        self.assertEqual(pager.label.back_color, (3, 6, 8))
+
+        # Select page 1
         proxy_1.push()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), True)
+        self.assertEqual(pager.enable_callback.enabled(action_2), False)
+
+        self.assertEqual(switch_proxy_1.color, (255, 255, 255))
+        self.assertEqual(switch_proxy_2.color, (3, 6, 8))
+        self.assertEqual(switch_proxy_1.brightness, 0.4)
+        self.assertEqual(switch_proxy_2.brightness, 0.3)
+
+        self.assertEqual(pager.label.text, "foo")
+        self.assertEqual(pager.label.back_color, (255, 255, 255))
+
+    
+    def test_proxy_pager_not_initialized(self):
+        pager = PagerAction(
+            pages = [
+                {
+                    "id": 1, 
+                    "color": (3, 4, 5),
+                    "text": "foo"
+                },
+                {
+                    "id": 2, 
+                    "color": (3, 6, 8),
+                    "text": "bar"
+                },
+            ],
+            led_brightness_on = 0.4,
+            led_brightness_off = 0.3
+        )
+
+        proxy_1 = pager.proxy(
+            page_id = 1
+        )
+
+        action_1 = MockAction({ "enableCallback": pager.enable_callback, "id": 1 })
+        action_2 = MockAction({ "enableCallback": pager.enable_callback, "id": 2 })
+
+        switch = MockFootswitch(
+            actions = [
+                action_1, 
+                action_2
+            ]
+        )
+        
+        switch_proxy_1 = MockFootswitch(
+            pixels = [0],
+            actions = [
+                proxy_1
+            ]
+        )
+
+        appl = MockController2(
+            inputs = [
+                switch, 
+                switch_proxy_1
+            ]
+        )
+        
+        proxy_1.init(appl, switch_proxy_1)
+
+        with self.assertRaises(Exception):
+            proxy_1.push()
+
+
+    def test_proxy_invalid_page(self):
+        pager = PagerAction(
+            pages = [
+                {
+                    "id": 1, 
+                    "color": (3, 4, 5),
+                    "text": "foo"
+                },
+                {
+                    "id": 2, 
+                    "color": (3, 6, 8),
+                    "text": "bar"
+                }
+            ],
+            select_page = 1
+        )
+
+        action_1 = MockAction({ "enableCallback": pager.enable_callback, "id": 1 })
+        action_2 = MockAction({ "enableCallback": pager.enable_callback, "id": 2 })
+
+        proxy = pager.proxy(
+            page_id = 10
+        )
+
+        switch = MockFootswitch(
+            actions = [
+                action_1, 
+                action_2
+            ]
+        )
+
+        switch_proxy = MockFootswitch(
+            actions = [
+                proxy
+            ]
+        )
+        
+        switch_pager = MockFootswitch(
+            pixels = [0],
+            actions = [
+                pager
+            ]
+        )
+        
+        appl = MockController2(
+            inputs = [
+                switch, 
+                switch_pager
+            ]
+        )
+        
+        pager.init(appl, switch_pager)
+        proxy.init(appl, switch_proxy)
+        
+        pager.update_displays()
+        proxy.update_displays()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), True)
+        self.assertEqual(pager.enable_callback.enabled(action_2), False)
+
+        pager.push()
+
+        self.assertEqual(pager.enable_callback.enabled(action_1), True)
+        self.assertEqual(pager.enable_callback.enabled(action_2), False)
+
+        proxy.push()
 
         self.assertEqual(pager.enable_callback.enabled(action_1), True)
         self.assertEqual(pager.enable_callback.enabled(action_2), False)