+++ web/htdocs/js/PySwitchRunner.js
@@ -17,6 +17,7 @@ class PySwitchRunner {
      *      updateIntervalMillis: 10,        Tick interval in milliseconds. On CircuitPython, the program does as much ticks as it can (in a while True loop),
      *                                       which in a browser woult block all user interaction, so the ticks are triggered in intervals. Mandatory.
      *      coverage: False                  Measure coverage
+     *      errorHandler: null               Optional error handler, providing a handle(exc) method
      * }
      */
     constructor(options, containerId) {
@@ -57,8 +58,6 @@ class PySwitchRunner {
         
         this.pyodide.FS.mkdir("parser/misc");
         await this.#loadModule("parser/misc/VisitorsWithStack.py", localPythonPath);
-        //await this.#loadModule("parser/misc/AddElementTransformer.py", localPythonPath);
-        // await this.#loadModule("parser/misc/RemoveDictElementTransformer.py", localPythonPath);
         await this.#loadModule("parser/misc/RemoveUnusedImportTransformer.py", localPythonPath);
         await this.#loadModule("parser/misc/CollectCommentsTransformer.py", localPythonPath);
         await this.#loadModule("parser/misc/AddImportsTransformer.py", localPythonPath);
@@ -70,21 +69,8 @@ class PySwitchRunner {
         await this.#loadModule("parser/misc/ImportExtractor.py", localPythonPath);
         await this.#loadModule("parser/misc/ReplaceAssignmentTransformer.py", localPythonPath);
         await this.#loadModule("parser/misc/AddAssignmentTransformer.py", localPythonPath);
-        //await this.#loadModule("parser/misc/ItemBase.py", localPythonPath);
-        // await this.#loadModule("parser/misc/Arguments.py", localPythonPath);
         await this.#loadModule("parser/misc/CodeExtractor.py", localPythonPath);
-        await this.#loadModule("parser/misc/CodeGenerator.py", localPythonPath);
-
-        // this.pyodide.FS.mkdir("parser/inputs");
-        // await this.#loadModule("parser/inputs/Action.py", localPythonPath);
-        // await this.#loadModule("parser/inputs/Actions.py", localPythonPath);
-        // await this.#loadModule("parser/inputs/Input.py", localPythonPath);
-        // await this.#loadModule("parser/inputs/InputReplacer.py", localPythonPath);
-        // await this.#loadModule("parser/inputs/InputAssignment.py", localPythonPath);
-        // await this.#loadModule("parser/inputs/CreateInputTransformer.py", localPythonPath);
-        
-        // this.pyodide.FS.mkdir("parser/display");
-        
+        await this.#loadModule("parser/misc/CodeGenerator.py", localPythonPath);        
 
         this.pyodide.FS.mkdir("wrappers");
         await this.#loadModule("wrappers/__init__.py", localPythonPath);
@@ -209,7 +195,11 @@ class PySwitchRunner {
         // Create external refs object (used to communicate with the python scripts)
         if (!window.externalRefs) {
             window.externalRefs = {};
-        }  
+            
+            if (this.#options.errorHandler) {
+                window.externalRefs.errorHandler = this.#options.errorHandler;
+            }            
+        }
         
         this.#initialized = true;
     }